Don't cull faces based on lighting
Doesn't make sense to do that unless there is at least a small bit of
light coming from the camera's "look from" position (because it causes
visible triangles to be culled).

Besides, need to drop this in order to move vertex lighting entirely to
the shader (which means that the per-face code won't even be able to see
lighting information).